A few small CSS changes to the search bar
While browsing Sugarizer I couldn't help but notice that the icons for the magnifying glass and cancel button were slightly out of alignment. I edited the CSS to increase the margin to the left of the magnifying glass from 5px to 6px, and to vertically center the magnifying glass, input text and cancel button in the search bar. This was done by removing the absolute top/left positions and replacing it with a percentage-based `top: 50%;`. However, this centers the top of the element meaning that the middle of the element is off-centre; to compensate for this I added `transform: translateY(-50%)` which would translate the element up by 50% of its height thus centering it vertically.