Add West Bank and Somaliland
The geojson files define the known countries in English.  The locale.ini file provides translations to other languages, and is also consulted in the case of English.  English is the only language activated in the Sugarizer version, so only necessary to add entries to the [en] and [*] sections.  The .po translation files in the repo tree do not seem to be used, and are not part of the original.