Delete playstore.png
This playstore.png file had no use in whole project, that's why I deleted it. And the same image is in the applestore.png file which has been used in the project's Readme section. If you still want that file I can rename it from "playstore.png" to "appstore.png" as it's a appstore's image not playstore's.